Fixes from open-to-all branch
