+++ client/src/Client/pages/Register/components/FormSteps/SummaryAndApproval/SummaryAndApproval.tsx
@@ -12,7 +12,7 @@ export const SummaryAndApproval = () => {
   return (
     <div className="flex flex-col gap-5 flex-grow">
       <p>{`${firstName} ${lastName},`}</p>
-      <p>נציגים של עזר לחיים יעיינו בפרטי ההרשמה שלך בשעות הקרובות על מנת לאשרה.</p>
+      <p>נציגים של עזר לחיים יעיינו בפרטי ההרשמה שלך על מנת לאשרה.</p>
       <p>האישור יגיע באמצעות הודעת SMS, אימייל או בפניה טלפונית.</p>
       <p className="self-end">
         רפואה שלמה,
