+++ client/src/Client/pages/Splash/Splash.tsx
@@ -3,7 +3,7 @@ import redisLogo from '../../../assets/redisLogo.svg';
 
 export const Splash = () => {
   return (
-    <div className="flex flex-col items-center justify-center w-screen h-screen fixed bg-white z-50 top-0 left-0">
+    <div className="flex flex-col items-center justify-center w-screen dvh-screen fixed bg-white z-50 top-0 left-0">
       <img src={logo} alt="logo" className="mb-2.5 animate-pulse" />
       <h1 className="text-center text-blue-500 font-medium w-64">
         שירות הסעות התנדבותי למבקרים בבתי חולים
