+++ client/src/index.css
@@ -13,3 +13,8 @@
         @apply text-blue-500 underline;
     }
 }
+
+.dvh-screen {
+    height: 100vh;
+    height: 100dvh;
+}
